Revert "Map for selecting location (#22)"
This reverts commit 2f48b00271b31f1f6b8c34c4852196b9d8b6ff29.
