+++ example.env
@@ -1,3 +1,2 @@
 REACT_APP_APIKEY=
-REACT_APP_MAPBOX_ACCESS_TOKEN=
 REACT_APP_API_NINJAS_API_KEY=
