+++ package.json
@@ -14,7 +14,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "mapbox-gl": "^2.5.1",
+    "dotenv": "^10.0.0",
     "bootstrap": "^5.1.3",
     "react": "^17.0.2",
     "react-bootstrap-typeahead": "^5.2.1",
@@ -47,6 +47,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {}
+  }
 }
