+++ src/components/PlacesTypeahead.js
@@ -1,45 +1,35 @@
-import { useContext, useEffect, useState } from "react";
+import { useState } from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { AsyncTypeahead } from "react-bootstrap-typeahead";
 import "react-bootstrap-typeahead/css/Typeahead.min.css";
-import { CityContext } from "../context/CityProvider";
 
 const LIMIT = 30;
 const SEARCH_URI = "https://api.api-ninjas.com/v1/city";
 
 export default function PlacesTypeahead(props) {
-  const { city, setCityData } = useContext(CityContext);
   const [isLoading, setIsLoading] = useState(false);
   const [options, setOptions] = useState([]);
-  const [selected, setSelected] = useState([city.name]);
+  const [selected, setSelected] = useState([]);
 
-  useEffect(() => {
-    setSelected([city.name]);
-  }, [city.name]);
-
-  const handleSearch = (query) => {
+  const handleSearch = query => {
     setIsLoading(true);
     fetch(`${SEARCH_URI}?name=${escape(query)}&limit=${LIMIT}`, {
       method: "GET",
       headers: {
         "Content-Type": "application/json",
-        "X-Api-Key": props.apiKey,
-      },
+        "X-Api-Key": props.apiKey
+      }
     })
-      .then((resp) => resp.json())
-      .then((data) => {
-        setIsLoading(false);
-        setOptions(
-          data && !data.error
-            ? data.map((city) => {
-                return city.name + ", " + city.country;
-              })
-            : []
-        );
-      });
+    .then(resp => resp.json())
+    .then(data => {
+      setIsLoading(false);
+      setOptions(data && !data.error ? data.map(city => {
+        return city.name + ", " + city.country;
+      }) : []);
+    });
   };
 
-  const handleChange = (selected) => {
+  const handleChange = selected => {
     setSelected(selected);
     props.onChange(selected);
   };
@@ -56,7 +46,6 @@ export default function PlacesTypeahead(props) {
       options={options}
       placeholder="Enter City Name..."
       isLoading={isLoading}
-      onKeyDown={props.onKeyDown}
     />
   );
 }
